eventsys: use uint32_t for events bitmask

+++ transport.c
@@ -84,7 +84,7 @@ int aura_get_pollfds(struct aura_node *node, const struct aura_pollfds **fds)
  * @param fd 
  * @param events 
  */
-void aura_add_pollfds(struct aura_node *node, int fd, short events)
+void aura_add_pollfds(struct aura_node *node, int fd, uint32_t events)
 {
 	struct aura_pollfds *ap;
